requirements.txt
python-dotenv>=1.0.0
pinecone>=5.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
openai>=1.0.0
